move signals to separate signals.py
