+++ custom_code/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class CustomCodeConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "custom_code"
+
+    def ready(self):
+        import custom_code.signals  # noqa: F401
