+++ custom_code/templates/custom_code/partials/target_classifications.html
@@ -5,9 +5,10 @@
 <div class="mb-2">
   {% with target.target_tags.all as target_tag_set %}
     {% for target_tag in target_tag_set %}
-      {% if target_tag.tag.name in 'auto classification ok,human classification ok,human classification unsure' %}
+      {% if target_tag.tag.name in 'auto classification ok,auto classification bad,human classification ok,human classification unsure' %}
         <span class="badge 
           {% if target_tag.tag.name == 'auto classification ok' %}badge-success
+          {% elif target_tag.tag.name == 'auto classification bad' %}badge-danger
           {% elif target_tag.tag.name == 'human classification ok' %}badge-primary
           {% elif target_tag.tag.name == 'human classification unsure' %}badge-warning
           {% endif %}">
