+++ custom_code/templates/custom_code/release_queue.html
@@ -67,6 +67,7 @@
         <th>Auto z</th>
         <th>Auto Prob</th>
         <th>Human Class</th>
+        <th>Human Subclass</th>
         <th>Human z</th>
         <th>Status</th>
       </tr>
